added to codeforces

+++ contest/224/a/Parallelepiped.java
@@ -3,9 +3,9 @@ import java.util.*;
 public class Parallelepiped {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int s1 = sc.nextInt();
-        int s2 = sc.nextInt();
-        int s3 = sc.nextInt();
+        long s1 = sc.nextLong();
+        long s2 = sc.nextLong();
+        long s3 = sc.nextLong();
         long abc = (long) Math.sqrt(s1 * s2 * s3);
         long edgeA = abc / s2;
         long edgeB = abc / s3;
